requirements.txt
# requirements.txt

google-cloud-speech==2.8.0
moviepy==1.0.3
numpy==1.21.2
opencv-python==4.5.3
sumy==0.8.1
yolo==v4